support entity in chain methods of dc
