+++ towhee/__init__.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-import threading
 from typing import List, Tuple, Union
 from towhee.dataframe import DataFrame
 from towhee.engine.engine import Engine, start_engine
@@ -20,10 +19,13 @@ from towhee.engine.pipeline import Pipeline
 from towhee.pipeline_format import OutputFormat
 from towhee.hub.file_manager import FileManagerConfig, FileManager
 from towhee.engine import register, resolve
-from towhee.engine.operator_loader import OperatorLoader
+from towhee.engine.factory import op, ops
 from towhee.hparam import param_scope, auto_param
 
-__all__ = ['DEFAULT_PIPELINES', 'pipeline', 'register', 'resolve', 'param_scope', 'auto_param', 'Build', 'Inject', 'dataset']
+__all__ = [
+    'DEFAULT_PIPELINES', 'pipeline', 'register', 'resolve', 'param_scope',
+    'auto_param', 'Build', 'Inject', 'dataset', 'op', 'ops'
+]
 
 DEFAULT_PIPELINES = {
     'image-embedding': 'towhee/image-embedding-resnet50',
@@ -45,6 +47,7 @@ class _PipelineWrapper:
         pipeline (`towhee.Pipeline`):
             Base `Pipeline` instance for which this object will provide a wrapper for.
     """
+
     def __init__(self, pipeline_: Pipeline):
         self._pipeline = pipeline_
 
@@ -72,7 +75,8 @@ class _PipelineWrapper:
         in_df = DataFrame('_in_df', cols)
         in_df.put(vargs)
         out_df = self._pipeline(in_df)
-        format_handler = OutputFormat.get_format_handler(self._pipeline.pipeline_type)
+        format_handler = OutputFormat.get_format_handler(
+            self._pipeline.pipeline_type)
         return format_handler(out_df)
 
     def __repr__(self) -> str:
@@ -120,32 +124,6 @@ def pipeline(pipeline_src: str, tag: str = 'main', install_reqs: bool = True):
     return _PipelineWrapper(pipeline_)
 
 
-def op(operator_src: str, tag: str = 'main', **kwargs):
-    """
-    Entry method which takes either operator tasks or paths to python files or class in notebook.
-    An `Operator` object is created with the init args(kwargs).
-    Args:
-        operator_src (`str`):
-            operator name or python file location or class in notebook.
-        tag (`str`):
-            Which tag to use for operators on hub, defaults to `main`.
-    Returns
-        (`typing.Any`)
-            The `Operator` output.
-    """
-    if isinstance(operator_src, type):
-        class_op = type('operator', (operator_src, ), kwargs)
-        return class_op.__new__(class_op, **kwargs)
-
-    loader = OperatorLoader()
-    if os.path.isfile(operator_src):
-        op_obj = loader.load_operator_from_path(operator_src, kwargs)
-    else:
-        op_obj = loader.load_operator(operator_src, kwargs, tag)
-
-    return op_obj
-
-
 def dataset(name: str, *args, **kwargs) -> 'TorchDataSet':
     """
     Get a dataset by name, and pass into the custom params.
@@ -170,7 +148,9 @@ def dataset(name: str, *args, **kwargs) -> 'TorchDataSet':
     from torchvision import datasets  # pylint: disable=import-outside-toplevel
     from towhee.data.dataset.dataset import TorchDataSet  # pylint: disable=import-outside-toplevel
     dataset_construct_map = {
-        'mnist': datasets.MNIST, 'cifar10': datasets.cifar.CIFAR10, 'fake': datasets.FakeData
+        'mnist': datasets.MNIST,
+        'cifar10': datasets.cifar.CIFAR10,
+        'fake': datasets.FakeData
         # 'imdb': IMDB  # ,()
     }
     torch_dataset = dataset_construct_map[name](*args, **kwargs)
@@ -200,6 +180,7 @@ class Build:
 
     >>> pipe = Build(template_variable_1='new_value').pipeline('pipeline_name')
     """
+
     def __init__(self, **kws) -> None:
         self._kws = kws
 
@@ -232,6 +213,7 @@ class Inject:
 
     and the value at the injection point is replace by the `Inject` API.
     """
+
     def __init__(self, **kws) -> None:
         self._injections = {}
         for k, v in kws.items():
@@ -241,84 +223,6 @@ class Inject:
         with param_scope() as hp:
             hp().injections = self._injections
             return pipeline(*arg, **kws)
-
-
-class _OperatorLazyWrapper:
-    """
-    operator wrapper for lazy initialization.
-    """
-    def __init__(self, name: str, index: Tuple[str], tag: str = 'main', **kws) -> None:
-        self._name = name.replace('.', '/').replace('_', '-')
-        self._index = index
-        self._tag = tag
-        self._kws = kws
-        self._op = None
-        self._lock = threading.Lock()
-
-    def __call__(self, *arg, **kws):
-        with self._lock:
-            if self._op is None:
-                self._op = op(self._name, self._tag, **self._kws)
-
-        if bool(self._index):
-            # Multi inputs.
-            if isinstance(self._index[0], tuple):
-                args = []
-                for i in self._index[0]:
-                    args.append(getattr(arg[0], i))
-                res = self._op(*args, **kws)
-            # Single input.
-            else:
-                args = getattr(arg[0], self._index[0])
-                res = self._op(args, **kws)
-
-            # Multi outputs.
-            if isinstance(res, tuple):
-                if not isinstance(self._index[1], tuple) or len(self._index[1]) != len(res):
-                    raise IndexError(f'Op has {len(res)} outputs, but {len(self._index[1])} indices are given.')
-                for i in range(len(res)):
-                    setattr(arg[0], self._index[1][i], res[i])
-                    arg[0].register(self._index[1][i])
-            # Single output.
-            else:
-                setattr(arg[0], self._index[1], res)
-                arg[0].register(self._index[1])
-
-            return arg[0]
-        else:
-            res = self._op(*arg, **kws)
-            return res
-
-    def train(self, *arg, **kws):
-        with self._lock:
-            if self._op is None:
-                self._op = op(self._name, self._tag, **self._kws)
-        return self._op.train(*arg, **kws)
-
-    @property
-    def function(self):
-        return self._name
-
-    @property
-    def init_args(self):
-        return self._kws
-
-    @staticmethod
-    def callback(name: str, index: Tuple[str], *arg, **kws):
-        if len(arg) == 0:
-            return _OperatorLazyWrapper(name, index, **kws)
-        else:
-            return _OperatorLazyWrapper(name, index, arg[0], **kws)
-
-
-ops = param_scope().callholder(_OperatorLazyWrapper.callback)
-"""
-Entry point for creating operator instances, for example:
-
->>> op_instance = ops.my_namespace.my_operator_name(init_arg1=xxx, init_arg2=xxx)
-
-An instance of `my_namespace`/`my_operator_name` is created.
-"""
 
 
 def _pipeline_callback(name, index, *arg, **kws):
